Don't use lockfile when checking for `is_broken`
Because a worker will hold the lock until it's finished, which will
block our `status()` function. This patch looks for the pidfile and
checks if the process is still alive.

Granted, there are a couple of race conditions, which result in an
exception or false positive for `is_broken` (not false negative though).
The UI will retry the API call, so it should be fine.